+++ lib/src/utils/color/material.dart
@@ -464,4 +464,3 @@ const amberA200 = const Color.rgb(255, 215, 64);
 const amberA400 = const Color.rgb(255, 196, 0);
 const amberA700 = const Color.rgb(255, 171, 0);
 const amber = amber500;
-